src/WeiboEmojiUpdateChecker/src/requirements.txt
onepush==1.1.1
python-dotenv==0.19.2